Change KcapiError 'code' to i32
The field 'code' in type KcapiError is an i64 which is not standard
for the underlying libkcapi. We need this to be an i32 so that we
can pass up errno style codes to any consumers of this error.

In the future, the 'code' and 'message' will be public fields, so
that the error returned by the kernel can be appropriately
interpreted by consumers.

This change makes treewide updates to accomodate this change.

+++ src/lib.rs
@@ -151,7 +151,7 @@ pub type KcapiResult<T> = std::result::Result<T, KcapiError>;
 ///
 #[derive(Debug, Clone)]
 pub struct KcapiError {
-    code: i64,
+    code: i32,
     message: String,
 }
 
@@ -225,7 +225,7 @@ impl IOVecTrait<Vec<u8>> for IOVec<Vec<u8>> {
     fn new(iov: Vec<Vec<u8>>) -> KcapiResult<Self> {
         if iov.is_empty() {
             return Err(KcapiError {
-                code: -libc::EINVAL as i64,
+                code: -libc::EINVAL,
                 message: format!(
                     "Cannot create an IOVec from a vector of length {}",
                     iov.len(),
